+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Examples/AlonzoBBODY.hs
@@ -30,6 +30,11 @@ import Cardano.Ledger.BaseTypes (
  )
 import Cardano.Ledger.Block (Block (..))
 import Cardano.Ledger.Coin (Coin (..))
+import Cardano.Ledger.Conway.Rules (ConwayCertsPredFailure (..), ConwayLedgerPredFailure (..))
+import qualified Cardano.Ledger.Conway.Rules as Conway (
+  ConwayBbodyPredFailure (..),
+  ConwayCertPredFailure (..),
+ )
 import Cardano.Ledger.Credential (
   Credential (..),
   StakeCredential,
@@ -127,7 +132,7 @@ tests =
     "Generic Tests, testing Alonzo PredicateFailures, in postAlonzo eras."
     [ alonzoBBODYexamplesP Alonzo
     , alonzoBBODYexamplesP Babbage
-    -- alonzoBBODYexamplesP Conway TODO
+    , alonzoBBODYexamplesP Conway
     ]
 
 alonzoBBODYexamplesP ::
@@ -681,7 +686,7 @@ testBBodyState pf =
 
 -- ============================== Helper functions ===============================
 
-makeTooBig :: Proof era -> AlonzoBbodyPredFailure era
+makeTooBig :: Proof era -> PredicateFailure (EraRule "BBODY" era)
 makeTooBig proof@Alonzo =
   ShelleyInAlonzoBbodyPredFailure
     . LedgersFailure
@@ -698,8 +703,13 @@ makeTooBig proof@Babbage =
     . DelplFailure
     . PoolFailure
     $ PoolMedataHashTooBig (coerceKeyRole . hashKey . vKey $ someKeys proof) (hashsize @Mock + 1)
--- makeTooBig proof@Conway =
---   ShelleyInAlonzoBbodyPredFailure . LedgersFailure . LedgerFailure . ConwayCertsFailure . CertFailure . PoolFailure $ ConwayPoolPredFailure -- FIXME: This needs fixing after POOL rules are implemented for Conway
+makeTooBig proof@Conway =
+  Conway.LedgersFailure
+    . LedgerFailure
+    . ConwayCertsFailure
+    . CertFailure
+    . Conway.PoolFailure
+    $ PoolMedataHashTooBig (coerceKeyRole . hashKey . vKey $ someKeys proof) (hashsize @Mock + 1)
 makeTooBig proof = error ("makeTooBig does not work in era " ++ show proof)
 
 coldKeys :: Crypto c => KeyPair 'BlockIssuer c