+++ eras/conway/impl/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 1.16.0.0
 
+* Add `maxRefScriptSizePerBlock` and `maxRefScriptSizePerTx` to `Cardano.Ledger.Conway.Rules`
+* Add `ConwayBBODY` and `ConwayBbodyPredFailure` type for BBody rule
 * Added `ConwayCommitteeIsUnknown` predicate failure to  `ConwayGovCertPredFailure`
 * Added `ceCurrentCommittee` and `ceCommitteeProposals` to `CertEnv`
 * Added `certsCurrentCommittee` and `certsCommitteeProposals` to `CertsEnv`